bello/requirements.txt
docx==0.2.4
langchain_google_genai==2.1.4
PyPDF2==3.0.1
python-dotenv==1.1.0
python_docx==1.1.2
